+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 - graphbrain.stats obsolete package.
 - graphbrain.cognition obsolete package.
 - graphbrain.op obsolete module.
+- Hyperedge .predicate() and .predicate_atom().
 
 ## [0.5.0] - 28-07-2021
 ### Added